backend/requirements.txt
fastapi
fastapi[standard]
uvicorn
supabase
pydantic-settings
python-dotenv
httpx
pytest
email-validator
google-adk>=1.20.0
ag-ui-adk
google-cloud-logging
google-cloud-aiplatform
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-gcp-trace
opentelemetry-exporter-gcp-monitoring
opentelemetry-instrumentation-fastapi
ag-ui-adk
opencv-python
chromadb
